flask-server/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
marshmallow==3.21.1
pytest==8.1.1
black==24.2.0
flake8==7.0.0
